rename mod transaction => transaction_stream
